zune: Merge new changes
* zune: Fix clippy errors.


* inflate: Change error type to allow errors to return decompressed data up until the error was encountered.

Addresses https://github.com/etemesi254/zune-image/issues/33


* png/errors: Fix zlib error


* core/lib: Remove `u16_to_u8` which is unsound

+++ crates/zune-jpeg/src/color_convert/avx.rs
@@ -415,5 +415,5 @@ unsafe fn clamp_avx(reg: __m256i) -> __m256i
 #[inline]
 const fn shuffle(z: i32, y: i32, x: i32, w: i32) -> i32
 {
-    ((z << 6) | (y << 4) | (x << 2) | w) as i32
+    (z << 6) | (y << 4) | (x << 2) | w
 }
